Bug 1867743 - Part 4: Introduce PopoverOpenedInMode r?#dom-core
`PopoverOpenedInMode` forces popovers to have a fixed state during their opened
lifetime, which may be different from the attribute state. They can use the
PopoverAttributeState as they can only be opened in modes that match their
attribute state.

This change introduces this new mode in the minimum ally viable way -
introducing it just for the check in HideAllPopoversUntil step 7; and the
callsites where it is set.

Differential Revision: https://phabricator.services.mozilla.com/D276370