Bug 1867743 - Part 6: Refactor popover methods to no longer hardcode to Auto mode r?#dom-core
The methods CloseEntirePopoverList, HidePopoverStackUntil, AutoPopoverList, and
IsAutoPopover all either take a popover list (of hint or auto) or have
equivalent methods (e.g. HintPopoverList). In order to introduce popover=hint,
this change refactors these methods to take a PopoverAttributeState which
determines the "opened in mode" to filter the popover list by, ensuring that
these methods can act on either Auto or Hint popovers.

This change has no behavioural changes - everything should function the
same - but it is preparatory work for popover=hint.

Differential Revision: https://phabricator.services.mozilla.com/D276372